requirements.txt
mattsollamatools
sentence-transformers
fastapi
python-multipart
uvicorn[standard]
protobuf
grpcio
soundfile
feedparser
newspaper3k
pypdf2
langdetect
langchain
docarray
#faiss-gpu
llama-index-core 
llama-index-llms-openai
llama-index-readers-file 
llama-index-llms-ollama 
llama-index-embeddings-huggingface
llama-index-embeddings-openai
llama-index-llms-openai_like
llama-index-vector-stores-milvus
requests_toolbelt
pandas 
openpyxl 
tabulate
python-docx
#pypandoc
mammoth 
pdfkit
docx2txt
pymupdf
python-pptx
markdown
pyautogen
#nvidia_ace-1.0.0-py3-none-any.whl
numba
# git+https://github.com/Good1521/graphrag.git@v0.5.0
PyJWT
passlib
bcrypt
fastapi_mail
fastapi_users_db_sqlalchemy
aiosqlite
pymongo
motor
pymilvus
trafilatura
databases
scrapy
pytest
cairosvg
huggingface-hub
